Update liquidation.md
I have a problem with following:

   To make sure that the required surplus of collateral exists at all times, a class of users called Keepers are charged with maintaining a constant watch for CDPs that become unsafe or under-collateralized. These Keepers are a special category of the Dai Stablecoin System users. They are the actors in the system who are incentivized to make sure that the outstanding Dai supply remains fully collateralized and solvent. They help maintain the health of the entire ecosystem by ensuring that unsafe CDPs are offered up for Liquidation as quickly as possible. This is particularly important during rapid market downturns as the collateral value could fail to satisfy debt obligations.

The answer for very basic question "Why does Liquidation exist?" is diverted into explanation of keepers role which is not really relevant here. Liquidation is needed for some reason (which is explained in 1st paragraph) and keepers are just a way to do liquidations, but there could other ways. So I'd suggest moving this to separate question ("Who triggers liquidation?") Also, this paragraph could be written in more concise way. 

"To make sure that the required surplus of collateral exists at all times, a class of users called Keepers are charged with maintaining a constant watch for CDPs that become unsafe or under-collateralized." says basically the same as "They help maintain the health of the entire ecosystem by ensuring that unsafe CDPs are offered up for Liquidation as quickly as possible."