Update governance.md
Please, please, ELI5 this bit more...

Instead of "the continuity of staked votes challenges and reinforces the status quo of the system through movements of the majority of votes between desired new proposals and the most recent successful proposal." (in 3rd paragraph), why not just say in *first* sentence, that Continuous Approval Voting means, that voters can leave the vote at the existing proposal or move it to a new one. Sorry I can't formulate it, but I think is should be more straightforward and simple.