What exactly is Stability Fee?
There's huge problem with this FAQ as a whole since it use the same term "Stability Fee" for two related, but distinct things.

One:

"The Stability Fee is a Risk Parameter designed to address imbalances in supply and demand for the Dai token which could result from periods of low or negative growth in the cryptocurrency markets."

"The Stability Fee is a rate that primarily affects the Dai supply since it alters the cost of creating Dai."

"The Risk Team will propose the thresholds for altering the Stability Fee"

Two:

"The Stability Fee is calculated _continuously_. It is denominated in Dai and can be paid in DAI or MKR."

"The Maker smart contracts collect a Stability Fee which is calculated against the total amount of DAI drawn against collateral held in a CDP. "

and 2in1

"This a variable rate fee which can change when MKR holders vote on proposals put forth by the Interim Risk Team at MakerDAO."

This leads to a lot of statements that contradict each other ("It is not possible to schedule or make predictions about when the fee should change or what the new rates could be."  vs "The Stability Fee is calculated _continuously_.")

I've always advocated to change the Risk Parameter name to "Stability Rate" and use Stability Fee only when talking about amount of DAI/MKR that user has to pay. Or any other DISTINCT names for these two distinct entities. It's impossible to continue the practice of mixing those two into one (except perhaps very informal communication).