Update stability-fee.md
Please consider if this should stay, in my opinion it's bit redundant, the effect of stability fee is discussed above:

"The Stability Fee is a rate that primarily affects the Dai supply since it alters the cost of creating Dai. The cheaper it is to borrow Dai, the more users are incentivized to do so. Conversely, when the fee is higher, fewer users will want to borrow Dai. MKR Token holders are able to set this rate to maintain the health of the peg."